aider: Corrected grammar in the `README.md` file.
# Aider chat conversation:

USER: correct grammar

ASSISTANT: To correct the grammar in the `README.md` file, I will provide a series of *SEARCH/REPLACE* blocks for each section that requires grammatical corrections. These corrections will address issues such as typographical errors, consistency in terminology, and clarity in descriptions.

### Correction 1: Introduction Section

README.md
<source>
<<<<<<< SEARCH
Reinforcement learning with human feedback (RLHF) on-chain, leverging on-chain trust, scalability, transparensy, and reputation.
=======
Reinforcement learning with human feedback (RLHF) on-chain, leveraging on-chain trust, scalability, transparency, and reputation.
>>>>>>> REPLACE
</source>

### Correction 2: Motivation Section

README.md
<source>
<<<<<<< SEARCH
Running this process deterministically on-chain allows to have trusted specialized alligned LLMs.
=======
Running this process deterministically on-chain allows us to have trusted, specialized, aligned LLMs.
>>>>>>> REPLACE
</source>

README.md
<source>
<<<<<<< SEARCH
This project also levereges well developed and decentralized network of nodes.
=======
This project also leverages a well-developed and decentralized network of nodes.
>>>>>>> REPLACE
</source>

### Correction 3: How it Works Section

README.md
<source>
<<<<<<< SEARCH
The LLM runs as an oprimistic rollup onto EVM chain.
=======
The LLM runs as an optimistic rollup onto the EVM chain.
>>>>>>> REPLACE
</source>

README.md
<source>
<<<<<<< SEARCH
These vouchers allow to validate and post the result of running the LLM on to the EVM; for that the vouchers need to be executed as an EVM transaction.
=======
These vouchers allow us to validate and post the result of running the LLM onto the EVM; for that, the vouchers need to be executed as an EVM transaction.
>>>>>>> REPLACE
</source>

### Correction 4: User Flow Section

README.md
<source>
<<<<<<< SEARCH
1. When LLM is done, you will see "Off-chain" reponses in the table. You will need to wait for the voucher proofs to be ready. Once the proofs are ready, you can **post** the LLM responses on to the chain by clicking "Post 0" and "Post 1".
=======
1. When LLM is done, you will see "Off-chain" responses in the table. You will need to wait for the voucher proofs to be ready. Once the proofs are ready, you can **post** the LLM responses onto the chain by clicking "Post 0" and "Post 1".
>>>>>>> REPLACE
</source>

### Correction 5: Deploy the Contract and Cartesi Section

README.md
<source>
<<<<<<< SEARCH
Clone this repo and its submodules that are repsonsible for backend and front-end.
=======
Clone this repo and its submodules that are responsible for backend and front-end.
>>>>>>> REPLACE
</source>

### Correction 6: Limitation Section

README.md
<source>
<<<<<<< SEARCH
- Cartesi currently does not suport vectorization. the inference is 10_000x slower than on CPU.
=======
- Cartesi currently does not support vectorization. The inference is 10,000x slower than on CPU.
>>>>>>> REPLACE
</source>

README.md
<source>
<<<<<<< SEARCH
- Currently, there is a problem running more than 90 tokens
=======
- Currently, there is a problem with running more than 90 tokens.
>>>>>>> REPLACE
</source>

These corrections will improve the readability and professionalism of the `README.md` document.